fix ruamel.yaml on 0.14.X for now
Thank you for using ruamel.yaml. 

There will be API changes in the 0.15+ versions, that might lead to warnings that 
your users could see. 
According to PyPI there is no fix on a version number because `setup.py` and `requirements.txt` are out of sync (and the setup.py info is used: ruamel.yaml>=0.11.11. If I install your package in a virtualenv with pip I get ruamel.yaml 0.15.0. 

With this change `install_requires`, is automatically read in `requirements.txt` (stripping comment lines), and `yoke` will use ==0.13.11 (you might want to change that to <0.15)